web/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
celery-redbeat==2.2.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
daphne==4.1.2
Deprecated==1.2.18
Django==5.1.9
django-allauth[socialaccount]==65.3.0
django-celery-beat==2.7.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.3
django-google-spanner==4.0.1
django-ipware==7.0.1
django-prometheus==2.3.1
django-redis==5.4.0
django-structlog==9.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
drf-spectacular==0.27.2
google-api-core==2.25.0
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-pubsub==2.28.0
google-cloud-spanner==3.55.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.73.0rc1
grpcio-status==1.71.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pillow==11.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
pytz==2025.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
sentry-sdk==2.22.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
structlog==25.4.0
tenacity==9.1.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.4.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.22.0
zope.interface==7.2
